requirements.txt
# This is a requirements file for a Python project.
yfinance==0.1.70
beautifulsoup4==4.12.2
matplotlib==3.5.2
Flask==2.3.2
pandas==2.0.3
numpy==1.26.2

# Machine learning
scikit-learn==1.3.0

#render HTML templates with Jinja2 (commonly used by Flask)
Jinja2==3.1.2

# Optional: if you use gunicorn for production (Heroku, etc.)
gunicorn==20.1.0


SQLAlchemy==1.4.46
psycopg2-binary==2.9.6
# Optional: if you're using environment variables
python-dotenv==0.21.0